requirements.txt
requests
python-dotenv
pandas
numpy
tenacity
